Added randomizer sample app
This app demonstrates job linking through the use of two components
that run as separate jobs. It is comprised of a simple Go REST API
and a PHP front-end.

When linked, the REST API generates a random number which the PHP
front-end displays along with information on both the Go and PHP
job instances.

Combined with job route balancing, this enables the randomizer app
to have varied output with each page refresh.